pyproject.toml
[project]
name = "pdfactory"
version = "0.1.0"
description = "Open-source service for automated PDF certificate generation from templates with API support."
authors = [{ name = "Artiom Cozlovici", email = "kozakovi4@outlook.com" }]
license = { file = "LICENSE" }
readme = "README.md"

requires-python = ">=3.14,<4.0"
dependencies = [
    "django (>=6.0,<7.0)",
    "django-unfold (>=0.73.1,<0.74.0)",
    "django-stubs[compatible-mypy] (>=5.2.8,<6.0.0)",
    "django-model-utils (>=5.0.0,<6.0.0)",
    "django-simple-history (>=3.11.0,<4.0.0)",
]


[dependency-groups]
dev = [
    "pylint (>=3.0,<4)",
    "django-stubs (>=5.2.8,<6.0.0)",
    "bandit (>=1.9.2,<2.0.0)",
    "mypy (>=1.19.0,<2.0.0)",
    "pylint-django (>=2.6.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false


[tool.mypy]
python_version = "3.12"
mypy_path = ["src"]
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = ["unfold.*", "simple_history.*"]
follow_untyped_imports = true


[tool.django-stubs]
django_settings_module = "main.settings"


[tool.ruff]
line-length = 120


[tool.pylint.main]
load-plugins = ["pylint_django", "pylint_django.checkers.migrations"]
django-settings-module = "main.settings"
ignore-paths = ["src/.*/migrations", "src/media", "src/static"]

[tool.pylint.format]
max-line-length = 120

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-few-public-methods",
]
